H2 R2DBC

+++ Reactive/fluxh2reactive/src/main/java/vn/techmaster/fluxh2reactive/repository/PostRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Flux;
 import vn.techmaster.fluxh2reactive.model.Post;
 
-public interface PostRepo extends ReactiveCrudRepository<Post, Integer>{
+public interface PostRepo extends ReactiveCrudRepository<Post, Long>{
   @Query("SELECT * FROM post WHERE title like $1")
   Flux<Post> findByTitleContains(String name);
 }
